+++ bank_system/settings.py
@@ -30,17 +30,27 @@ ALLOWED_HOSTS = []
 
 # Application definition
 
-INSTALLED_APPS = [
+BUILT_IN_APPS = [
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'banks',
+]
+
+# Third-party apps
+THIRD_PARTY_APPS = [
     'rest_framework',
     'rest_framework.authtoken',
 ]
+
+# Custom apps
+CUSTOM_APPS = [
+    'banks',
+]
+
+INSTALLED_APPS = BUILT_IN_APPS + THIRD_PARTY_APPS + CUSTOM_APPS
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
